# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Golf Pool App API
  version: 1.0.0
  description: |
    API for managing golf pools, teams, tournaments, and leaderboards.

    All endpoints (except `/health`) require a Firebase Auth ID token passed as a Bearer token.
    Admin-only endpoints additionally require the authenticated user to have `admin: true` in Firestore.

    Make sure to use the correct server URL depending on your environment:
    - Local development: `http://127.0.0.1:5001/golf-pool-app-492300/us-central1/api`
    - Production: `https://us-central1-golf-pool-app-492300.cloudfunctions.net/api`
servers:
  - url: http://127.0.0.1:5001/golf-pool-app-492300/us-central1/api
    description: Local emulator
  - url: https://us-central1-golf-pool-app-492300.cloudfunctions.net/api
    description: Production (Firebase Cloud Function)s
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Firebase Auth ID token
  schemas:
...
    PlayerInput:
      type: object
      required: [name, odds]
      properties:
        name:
          type: string
          example: "Scottie Scheffler"
        odds:
          type: string
          example: "+450"
    PlayerDetail:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        odds:
          type: string
        espnMapped:
          type: boolean
...
    TournamentSummary:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        espnEventId:
          type: string
          nullable: true
        startDate:
          type: string
          format: date-time
          nullable: true
        endDate:
          type: string
          format: date-time
          nullable: true
        status:
          type: string
          enum: [upcoming, active, completed]
...
    Error:
      type: object
      properties:
        error:
          type: string
      example:
        error: "Resource not found"
    Success:
      type: object
      properties:
        success:
          type: boolean
          example: true
  responses:
    BadRequest:
      description: Invalid request body or parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing or invalid auth token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Authenticated but not authorized (e.g. not admin, not team owner)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    BadGateway:
      description: Upstream ESPN API error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    SuccessResponse:
      description: Operation succeeded
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Success"
security:
  - BearerAuth: []
...
paths:
...
  /tournaments:
    get:
      summary: List tournaments
      description: Returns tournaments visible to the caller, sorted by start date ascending. Any signed-in user.
      tags: [Tournaments]
      parameters:
        - in: query
          name: status
          required: false
          schema:
            type: string
            enum: [upcoming, active, completed]
          description: Optional status filter
      responses:
        "200":
          description: Tournament list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TournamentSummary"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      summary: Create a tournament
      description: "**Admin only.** Creates a new tournament."
      tags: [Tournaments]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, startDate, endDate]
              properties:
                name:
                  type: string
                  example: "The Masters 2026"
                espnEventId:
                  type: string
                  nullable: true
                  description: ESPN event ID for score syncing (optional)
                  example: "401353232"
                startDate:
                  type: string
                  format: date-time
                  example: "2026-04-10T13:00:00Z"
                endDate:
                  type: string
                  format: date-time
                  example: "2026-04-13T20:00:00Z"
      responses:
        "201":
          description: Tournament created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: "abc123"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /tournaments/{tournamentId}/players:
    get:
      summary: List players for a tournament
      description: Returns the player roster (name, odds, espnMapped) sorted by name. Any signed-in user; required by the team picker.
      tags: [Tournaments]
      parameters:
        - in: path
          name: tournamentId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Player roster
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlayerDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      summary: Bulk add players to a tournament
      description: "**Admin only.** Adds one or more players with odds to a tournament."
      tags: [Tournaments]
      parameters:
        - in: path
          name: tournamentId
          required: true
          schema:
            type: string
          example: "abc123"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [players]
              properties:
                players:
                  type: array
                  minItems: 1
                  items:
                    $ref: "#/components/schemas/PlayerInput"
            example:
              players:
                - name: "Scottie Scheffler"
                  odds: "+450"
                - name: "Rory McIlroy"
                  odds: "+600"
      responses:
        "201":
          description: Players added successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  playerIds:
                    type: array
                    items:
                      type: string
                  count:
                    type: integer
              example:
                playerIds: ["p1", "p2"]
                count: 2
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /tournaments/{tournamentId}/sync-espn:
    post:
      summary: Auto-sync unmapped players to ESPN IDs
      description: |
        **Admin only.** Fetches the ESPN scoreboard and fuzzy-matches our player names to ESPN competitor IDs.
        Only processes players where `espnMapped` is `false`.
      tags: [Tournaments]
      parameters:
        - in: path
          name: tournamentId
          required: true
          schema:
            type: string
          example: "abc123"
      responses:
        "200":
          description: Sync result summary
          content:
            application/json:
              schema:
                type: object
                properties:
                  matched:
                    type: integer
                  unmatched:
                    type: integer
                  matchedPlayers:
                    type: array
                    items:
                      type: object
                      properties:
                        playerId:
                          type: string
                        espnId:
                          type: string
                        name:
                          type: string
                  unmatchedPlayers:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        name:
                          type: string
              example:
                matched: 48
                unmatched: 2
                matchedPlayers: []
                unmatchedPlayers: []
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "502":
          $ref: "#/components/responses/BadGateway"
  /tournaments/players/{playerId}/espn-link:
    put:
      summary: Manually link a player to an ESPN ID
      description: "**Admin only.** Used to fix players that failed auto-sync."
      tags: [Tournaments]
      parameters:
        - in: path
          name: playerId
          required: true
          schema:
            type: string
          example: "player_xyz"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [espnId]
              properties:
                espnId:
                  type: string
                  example: "3448"
      responses:
        "200":
          $ref: "#/components/responses/SuccessResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ─────────────────────────────────────────────
  # Pools
  # ─────────────────────────────────────────────
...